Remove Util class and move to StackdriverExportUtils.
